refactor: domain들을 한 단계 더 패키징함
