+++ src/main/java/com/backend/backend/controller/post/PostController.java
@@ -3,7 +3,7 @@ package com.backend.backend.controller.post;
 import com.backend.backend.controller.post.Dto.PostModifyRequest;
 import com.backend.backend.controller.post.Dto.PostReadResponse;
 import com.backend.backend.controller.post.Dto.PostWriteRequest;
-import com.backend.backend.domain.Post;
+import com.backend.backend.domain.post.Post;
 import com.backend.backend.service.PostService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
