+++ src/main/java/com/backend/backend/controller/review/ReviewController.java
@@ -2,7 +2,7 @@ package com.backend.backend.controller.review;
 
 import com.backend.backend.controller.review.dto.ReviewReadResponse;
 import com.backend.backend.controller.review.dto.ReviewWriteRequest;
-import com.backend.backend.domain.Review;
+import com.backend.backend.domain.review.Review;
 import com.backend.backend.service.ReviewService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
