+++ src/main/java/com/backend/backend/domain/review/Review.java
@@ -1,51 +1,48 @@
-package com.backend.backend.domain;
+package com.backend.backend.domain.review;
 
+import com.backend.backend.domain.member.Member;
+import com.backend.backend.domain.post.Post;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
 @EntityListeners(AuditingEntityListener.class)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Post {
+public class Review {
 
     @Id
-    @Column(name = "post_id")
     @GeneratedValue
+    @Column(name = "review_id")
     private Long id;
 
     @JsonIgnore
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id",nullable = false)
-    Member writer;
+    @JoinColumn(name = "post_id", nullable = false)
+    private Post post;
 
-    @Column(nullable = false)
-    @Setter
-    private String postName;
-    @Setter
-    @Lob
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id",nullable = false)
+    private Member writer;
     private String content;
 
     @CreatedDate
     @LastModifiedDate
     private LocalDateTime date;
 
-    @OneToMany(mappedBy = "post", cascade = CascadeType.ALL)
-    @OrderBy("date desc")
-    private List<Review> reviews = new ArrayList<>();
-
     @Builder
-    public Post(Member writer, String postName, String content) {
+    public Review(Post post, Member writer, String content) {
+        this.post = post;
         this.writer = writer;
-        this.postName = postName;
         this.content = content;
     }
 }
