+++ src/main/java/com/backend/backend/domain/transactionDetail/TransactionStatus.java
@@ -1,4 +1,4 @@
-package com.backend.backend.domain;
+package com.backend.backend.domain.transactionDetail;
 
 public enum TransactionStatus {
     TRADING,CANCEL,COMPLETE
