+++ src/main/java/com/backend/backend/repository/memberRepository/MemberRepository.java
@@ -1,10 +1,8 @@
 package com.backend.backend.repository.memberRepository;
 
-import com.backend.backend.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
+import com.backend.backend.domain.member.Member;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface MemberRepository{
     /**
