+++ src/main/java/com/backend/backend/repository/postRepository/MysqlPostRepository.java
@@ -1,6 +1,6 @@
 package com.backend.backend.repository.postRepository;
 
-import com.backend.backend.domain.Post;
+import com.backend.backend.domain.post.Post;
 import com.backend.backend.repository.OrderBy;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
