+++ src/test/java/com/backend/backend/controller/review/ReviewControllerTest.java
@@ -1,9 +1,9 @@
 package com.backend.backend.controller.review;
 
 import com.backend.backend.controller.review.dto.ReviewWriteRequest;
-import com.backend.backend.domain.Member;
-import com.backend.backend.domain.Post;
-import com.backend.backend.domain.Review;
+import com.backend.backend.domain.member.Member;
+import com.backend.backend.domain.post.Post;
+import com.backend.backend.domain.review.Review;
 import com.backend.backend.repository.memberRepository.MemberRepository;
 import com.backend.backend.repository.postRepository.PostRepository;
 import com.backend.backend.repository.reviewRepository.ReviewRepository;
